Update go.yml to add OWAP ZAP Baseline Scan
Given that the action spins up a docker container in an ubuntu virtual machine, the zap scan (which also resides in a docker container) should be able to access the app within the docker container via localhost. I could be wrong in that assumption, however. We could also have it run the scan on the deployed app within cloud.gov and point it to that URL. Not sure what is best, but I want to try this and see if it can reach it via localhost within the virtual env.