Test OWASP ZAP
Turns out by the nature of github actions, communications between docker containers (in our case the app and the zap action) is not ok. Also, the actions are run sequentially, so after the build and test, the container is brought down. However, I still think that pointing zap at a container in this environment before exposing it to the public in the gov cloud may be the best option.